# game-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Strategy game backend with users, villages, resources, buildings, and units.",
        "title": "Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/buildings": {
            "get": {
                "description": "Zwraca listę budynków i ich poziomów w wybranej wiosce",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buildings"
                ],
                "summary": "Pobierz budynki wioski",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID wioski",
                        "name": "village_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing or invalid village_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/buildings/cost": {
            "get": {
                "description": "Zwraca koszt następnego poziomu dla podanego budynku",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buildings"
                ],
                "summary": "Pobierz koszt ulepszenia budynku",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID wioski",
                        "name": "village_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Typ budynku (np. lumbermill, claypit, ironmine, warehouse, barracks)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing params or invalid village_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Building not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/buildings/upgrade": {
            "put": {
                "description": "Podnosi poziom budynku, jeśli gracza stać na koszt",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buildings"
                ],
                "summary": "Ulepsz budynek",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID wioski",
                        "name": "village_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Typ budynku (np. lumbermill, claypit, ironmine, warehouse, barracks)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing params or invalid village_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Not enough resources",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Building not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Zwraca token JWT po poprawnym zalogowaniu",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logowanie użytkownika",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nazwa użytkownika",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hasło",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Tworzy nowego użytkownika wraz ze startową wioską, zasobami, budynkami i jednostkami",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Rejestracja użytkownika",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nazwa użytkownika",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Adres e-mail",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hasło (min 8 znaków, wielka, mała, cyfra, znak specjalny)",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "User exists or DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}